requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
cssselect2==0.8.0
daphne==4.2.0
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-crum==0.7.9
django-filter==25.1
django-health-check==3.19.2
django-jazzmin==3.0.1
django-mfa2==3.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
ecdsa==0.19.1
et_xmlfile==2.0.0
Faker==37.4.0
fido2==1.1.3
fonttools==4.58.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
msgpack==1.1.1
openpyxl==3.1.5
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
polars==1.30.0
polars-lts-cpu==1.30.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyotp==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-u2flib-server==5.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.4.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1